refactor(timestamp): initial refactor for A2A SDK v1.0
