+++ samples/python/extensions/timestamp/pyproject.toml
@@ -16,6 +16,7 @@ dev = [
     "httpx>=0.28",
     "pytest>=8.0",
     "pytest-asyncio>=0.23",
+    "uvicorn>=0.34.2",
 ]
 
 [tool.pytest.ini_options]
